backend-django/requirements.txt
django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.4
reportlab==4.1.0
python-dotenv==1.0.1
Pillow==10.3.0
